Fix tests by moving packages
