+++ p2p-sync/src/test/java/org/smartregister/p2p/sync/handler/SyncSenderHandlerTest.java
@@ -1,4 +1,4 @@
-package org.smartregister.p2p.sync;
+package org.smartregister.p2p.sync.handler;
 
 import com.google.android.gms.nearby.connection.Payload;
 import com.google.android.gms.nearby.connection.PayloadTransferUpdate;
